+++ src/index-web.ts
@@ -9,7 +9,10 @@ export * from './chain/tezos/TezosLanguageUtil';
 export * from './chain/tezos/TezosMessageUtil';
 export * from './chain/tezos/TezosNodeReader';
 export * from './chain/tezos/TezosNodeWriter';
-export * from './chain/tezos/TezosProtocolHelper';
+export * from './chain/tezos/contracts/BabylonDelegationHelper';
+export * from './chain/tezos/contracts/DexterTokenHelper';
+export * from './chain/tezos/contracts/Tzip7ReferenceTokenHelper';
+export * from './chain/tezos/contracts/TCFBakerRegistryHelper';
 
 export * from './identity/tezos/TezosWalletUtil';
 export * from './identity/tezos/TezosFileWallet';